Separated each ObjectOID Parser into their own classes.
